+++ elmyra.ip.access.epo/CHANGES.rst
@@ -4,6 +4,7 @@ elmyra.ip.access.epo CHANGES
 
 development
 ===========
+- ui, middleware: propagate metadata of signed opaque parameter token downstream to show link expiry time (ttl) in user interface
 
 0.24.0
 ======